fix   [ ] [Profissional] Ao confirmar/enviar uma resposta de serviço, da crash

+++ SOMEI/SOMEI/Model/Structures/OrcamentoAnswerStruct.swift
@@ -12,7 +12,7 @@ struct OrcamentoAnswerStruct:Codable {
     
     var agendaId:Double
     var servicoId:Double
-    var valor:Double
+    var valor:Int
     var observacao:String
 
 }
